+++ src/components/Modals/CreateNewDatabase/CreateDBModal.jsx
@@ -1,17 +1,8 @@
 import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { useSetAtom, useAtomValue, useAtom } from "jotai";
+import { useSetAtom, useAtom } from "jotai";
 
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import fetchData from "../../../utils/axios";
-import {
-  currentDBIdAtom,
-  isListViewAtom,
-  currentDBNameAtom,
-  userAtom,
-  showCreateDBModalAtom,
-  dbFieldsAtom,
-} from "../../../atoms/atoms";
+import { showCreateDBModalAtom, dbFieldsAtom } from "../../../atoms/atoms";
+import useMutatePostDB from "../../../apis/useMutatePostDB";
 
 import Modal from "../../shared/Modal";
 import ContentWrapper from "../SharedItems/ContentWrapper";
@@ -23,15 +14,11 @@ import InputsArea from "../SharedItems/InputsArea";
 import InputWrapper from "../SharedItems/InputWrapper";
 import CreateDBInputList from "./CreateDBInputList";
 import Button from "../../shared/Button";
-import Loading from "../../shared/Loading";
 
 import CONSTANT from "../../../constants/constant";
 
 function CreateDBModal() {
-  const queryClient = useQueryClient();
-  const navigate = useNavigate();
-  const { userId } = useAtomValue(userAtom);
-
+  const fetchNewDatabase = useMutatePostDB();
   const [dbName, setdbName] = useState(null);
   const [fields, setFields] = useAtom(dbFieldsAtom);
 
@@ -39,9 +26,6 @@ function CreateDBModal() {
   const [isFieldNameEmpty, setIsFieldNameEmpty] = useState(false);
   const [isFieldNameDuplicate, setIsFieldNameDuplicate] = useState(false);
 
-  const setIsListView = useSetAtom(isListViewAtom);
-  const setCurrentDBId = useSetAtom(currentDBIdAtom);
-  const setCurrentDBName = useSetAtom(currentDBNameAtom);
   const setShowCreateDBModal = useSetAtom(showCreateDBModalAtom);
 
   function updateFieldName(index, event) {
@@ -65,33 +49,6 @@ function CreateDBModal() {
       },
     ]);
   }
-
-  async function fetchDatabase(newDatabase) {
-    const response = await fetchData(
-      "POST",
-      `/users/${userId}/databases`,
-      newDatabase,
-    );
-
-    return response;
-  }
-
-  const { mutate: fetchDatabaseSave, isLoading } = useMutation(fetchDatabase, {
-    onSuccess: result => {
-      setCurrentDBId(result.data.newDatabase._id);
-      setCurrentDBName(result.data.newDatabase.name);
-      setIsListView(true);
-
-      queryClient.refetchQueries(["userDbList"]);
-
-      navigate("/dashboard/listview");
-
-      setShowCreateDBModal(false);
-    },
-    onFailure: () => {
-      console.log("sending user to errorpage");
-    },
-  });
 
   function handleClickSave() {
     const names = [];
@@ -128,7 +85,7 @@ function CreateDBModal() {
         fields,
       };
 
-      fetchDatabaseSave(newDatabase);
+      fetchNewDatabase(newDatabase);
     }
   }
 
@@ -138,10 +95,6 @@ function CreateDBModal() {
     setIsFieldNameDuplicate(false);
 
     setdbName(event.target.value);
-  }
-
-  if (isLoading) {
-    return <Loading />;
   }
 
   return (
